+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
 ### Changed
 - There have been some internal changes to how styles are read from the app/addon trees. This should be a non-breaking change (and fixes a number of oddities, including previous incompatibility with lazy engines), but is enough of a shift to warrant a version bump and a note.
+- Attempting to import a nonexistent `@value` or `compose` a nonexistent class is now a hard error at build time.
+
+### Removed
+- Support for the deprecated `styles` property has been removed.
 
 ## 0.6.5 Time Marches On (March 16, 2017)
 ### Fixed